requirements.txt
# MLOps Assignment 2 - Cats vs Dogs Classification
# Requirements file with pinned versions for reproducibility

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Experiment Tracking
mlflow>=2.9.0

# Data Version Control
dvc>=3.30.0

# API Framework (for M2)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP requests for dataset download
requests>=2.31.0
kaggle>=1.5.0
